+++ panel/class-panel-restrictions.php
@@ -5,13 +5,13 @@ class ProfessionalDirectory_Admin {
     // Construtor da classe
     public function __construct() {
         add_action('pre_get_posts', [$this, 'restrict_service_view']);
-        add_filter('views_edit-professional_service', [$this, 'adjust_service_views']);
+        add_filter('views_edit-pdr_service', [$this, 'adjust_service_views']);
     }
 
     // Função para restringir a visualização de serviços no admin
     public function restrict_service_view($query) {
-        if (is_admin() && $query->is_main_query() && $query->get('post_type') === 'professional_service') {
-            if (current_user_can('professional') && !current_user_can('administrator')) {
+        if (is_admin() && $query->is_main_query() && $query->get('post_type') === 'pdr_service') {
+            if (current_user_can('service_provider') && !current_user_can('administrator')) {
                 $query->set('author', get_current_user_id());
             }
         }
@@ -19,7 +19,7 @@ class ProfessionalDirectory_Admin {
 
     // Função para ajustar as abas na tela de listagem de serviços
     public function adjust_service_views($views) {
-        if (get_post_type() === 'professional_service' && current_user_can('professional') && !current_user_can('administrator')) {
+        if (get_post_type() === 'pdr_service' && current_user_can('service_provider') && !current_user_can('administrator')) {
             // Remover a aba "All"
             if (isset($views['all'])) {
                 unset($views['all']);