+++ panel/panel-general-customizations.php
@@ -5,10 +5,10 @@
 function pdr_adjust_dashboard_menu() {
     global $submenu;
 
-    // Checa se o usuário atual é 'professional'
-    if (current_user_can('professional')) {
+    // Checa se o usuário atual é 'service_provider'
+    if (current_user_can('service_provider')) {
         // Muda a URL do Dashboard para a página do dashboard do professional
-        $submenu['index.php'][0][2] = 'admin.php?page=pdr-professional-dashboard';
+        $submenu['index.php'][0][2] = 'admin.php?page=pdr-service-provider-dashboard';
         // Remove o submenu "Home" indesejado
         unset($submenu['index.php'][0]);
     }
@@ -21,9 +21,9 @@ function pdr_redirect_dashboard() {
         $screen = get_current_screen();
         // Verifica se a tela atual é o dashboard
         if ($screen->id === "dashboard") {
-            // Redireciona apenas se o usuário for 'professional'
-            if (current_user_can('professional')) {
-                wp_redirect(admin_url('admin.php?page=pdr-professional-dashboard'));
+            // Redireciona apenas se o usuário for 'service_provider'
+            if (current_user_can('service_provider')) {
+                wp_redirect(admin_url('admin.php?page=pdr-service-provider-dashboard'));
                 exit;
             }
             // Não redireciona administradores, permitindo acesso ao dashboard padrão
@@ -34,7 +34,7 @@ add_action('current_screen', 'pdr_redirect_dashboard');
 
 
 function pdrEnqueueCustomAdminStyle() {
-    if (current_user_can('professional')) {
+    if (current_user_can('service_provider')) {
         wp_enqueue_style('pdr-custom-admin-style', plugin_dir_url(__FILE__) . '../panel/css/admin-customizations.css');
     }
 }